attrs/requirements.txt
attrs==23.2.0